Split up src/web/page into something more manageable
